requirements.txt
selenium
python-telegram-bot
Pillow